Unify three other features into a single file
